refactor(dynamodb_task_repository): improve task retrieval and deletion methods

+++ app/infrastructure/db/dynamodb_task_repository.py
@@ -2,6 +2,7 @@ import os
 from datetime import datetime
 
 import boto3
+from boto3.dynamodb.conditions import Key
 from loguru import logger
 
 from ...domain.task import Task, TaskDescription, TaskId, TaskStatus, TaskTitle
@@ -63,43 +64,50 @@ class DynamoDBTaskRepository(TaskRepository):
         task_id: TaskId,
     ) -> Task | None:
         """Find a task by its ID."""
-        # use gsi1pk
-        resp = self._table.get_item(
-            Key={
-                "GSI1PK": f"TASK#{task_id}",
-                "GSI1SK": f"TASK#{task_id}",
-            }
+        logger.info(f"Finding task by ID: {task_id}")
+
+        resp = self._table.query(
+            IndexName="GSI1",
+            KeyConditionExpression=Key("GSI1PK").eq(f"TASK#{task_id}")
+            & Key("GSI1SK").eq(f"TASK#{task_id}"),
         )
-        item = resp.get("Item")
-        if item is None:
+        items = resp["Items"]
+
+        if not items:
             return None
 
         return Task(
-            id=TaskId(str(item["id"])),
-            task_list_id=TaskListId(str(item["task_list_id"])),
-            title=TaskTitle(str(item["title"])),
-            description=TaskDescription(str(item["description"])),
-            status=TaskStatus(str(item["status"])),
-            created_at=datetime.fromisoformat(str(item["created_at"])),
+            id=TaskId(str(items[0]["task_id"])),
+            task_list_id=TaskListId(str(items[0]["task_list_id"])),
+            title=TaskTitle(str(items[0]["title"])),
+            description=TaskDescription(str(items[0]["description"])),
+            status=TaskStatus(str(items[0]["status"])),
+            created_at=datetime.fromisoformat(str(items[0]["created_at"])),
         )
 
     def delete(self, task_id: TaskId) -> None:
         """Delete a task by its ID."""
+        resp = self._table.query(
+            IndexName="GSI1",
+            KeyConditionExpression=Key("GSI1PK").eq(f"TASK#{task_id}"),
+        )
+        items = resp.get("Items", [])
+
+        if not items:
+            return
+
         self._table.delete_item(
             Key={
-                "GSI1PK": f"TASK#{task_id}",
-                "GSI1SK": f"TASK#{task_id}",
+                "PK": items[0]["PK"],
+                "SK": items[0]["SK"],
             }
         )
 
     def list_all(self, task_list_id: TaskListId) -> list[Task]:
         """List all tasks in the repository."""
         resp = self._table.query(
-            KeyConditionExpression="PK = :pk AND begins_with(SK, :sk)",
-            ExpressionAttributeValues={
-                ":pk": f"TASK_LIST#{task_list_id}",
-                ":sk": "TASK#",
-            },
+            KeyConditionExpression=Key("PK").eq(f"TASK_LIST#{task_list_id}")
+            & Key("SK").begins_with("TASK#"),
         )
 
         items = resp["Items"]
